Find Best Match in A Vector of Strings using Automatic Gap Detection

Find Best Match in A Vector of Strings using Automatic Gap Detection

Usage

find_best_match(pattern, strings, index = F, silent = F)

Arguments

pattern

A pattern to look for.

strings

A vector of strings to look for the pattern.

index

Logical. Whether to return index of the best matching item in the string vector (TRUE) or return the best matching item value (FALSE).

silent

Logical. Set to TRUE to disable result printing.

Author

Zhehao Hu